requirements.txt
label-studio
ultralytics
fastapi
uvicorn
opencv-python-headless
python-multipart
Pillow